Consistency check (#36)
* Added bentham encoding

* added yaml files

* updated fix_bentham notebook

* Change format of correct answer key in yaml files

* Updated a lot of code file

* Update patch_question_identifiers

* Updated all yaml files

* Fixed missing bentham encodings; added check_coverager.ipynb

* Moved fix notebooks in separate directory

* Added additional information to asserts in load_yaml

* Manually updated yaml file with missing information

* Created yaml files for Wiley, which I missed before

* updated check_inconsistencies

* updated check coverage

* Fixed Nature

* updated fix_inconsistencies

* Update rdp.csv, added additional naming conventions to data_processing

* modified the structure

* Run pre-commit

* Update docs and type hints for question.py

* Update docs and type hints in publisher.py

* Update docs and type hints in journal.py

* Update docs and type hints in csv_handler.py

* Update docs and type hints in yaml_handler.py

* Update docs and type hints in grouper.py

* Fix typo

* Add question_id to Question class

* ACS consistency checks

manually fixed data in wrong columns in rdp.csv;
added ACS yaml files;
added work flow to create ACS yaml files

* removed build folder

* add more journals to normalize journal function, add basic unit-test for data processing normalize_journal

* fix pre-commit again

* add README for dimensions data

* Minor aesthetic changes

* Move metadata outside of the package

* Update question types (use question id)

* update to analysis_example notebook

* added ieee yaml file

* updated rdp.csv

* updated rdp.csv

* Updated rdp.csv and added rdp.xlsx

Since some of our descriptions in the rdp.csv file have commas, the csv format may mess this up. Ich changed the the format so that there are quotes delimiting text. For safety, I also added a regular excel file.

* fixed Pleiades publishing

* added yaml files for MDPI

* some linting in normalize

* linting

* linting

* linting

* more linting

* added analysis workflows

* fix linting

* fix normalize.. to _normalize.. function names and data_processing_test

* fix union type operator for older python versions

* add missing import from typing import Optional

* formatting of notebooks and plots

---------

+++ data/processed/patch_question_identifiers.py
@@ -1,3 +1,4 @@
+import os
 import shutil
 from glob import glob
 
@@ -31,8 +32,21 @@ for pub in all_pubs:
     for j in all_journals:
         # create a backup copy of the original file
         j_backup = j + "_backup"
-        shutil.copy(j, j_backup)
+
+        if not os.path.isfile(j_backup):
+            shutil.copy(j, j_backup)
+        else:
+            raise "Backup files already exist."
+
         y = yaml.safe_load(open(j))
-        new_y = {question_number_to_question_id[i]: values for i, values in y.items()}
-        with open(j, "w") as f:
-            yaml.dump(new_y, f, sort_keys=False)
+
+        new_y = {}
+        update = False
+        for i, values in y.items():
+            if i in question_number_to_question_id:
+                update = True
+                new_y[question_number_to_question_id[i]] = values
+
+        if update:
+            with open(j, "w") as f:
+                yaml.dump(new_y, f, sort_keys=False)
